Refactor time-series charts into feature-based architecture
Reorganized time-series chart functionality from scattered locations in data-tracking into a cohesive analysis/time-series feature directory. Eliminated facade pattern in chart-data.ts by extracting misplaced logic and establishing explicit dependencies. Moved navigation logic from analysis to navigation feature for proper domain separation.

This migration restores feature boundaries, reduces indirection, and improves code discoverability by colocating related chart components, hooks, types, and logic in their correct domains.
