+++ src/features/analysis/time-series/time-series-chart.tsx
@@ -2,10 +2,12 @@ import { useState, useMemo, useEffect } from 'react'
 import { Area, AreaChart, XAxis, YAxis } from 'recharts'
 import { ChartContainer, ChartTooltip, ChartTooltipContent } from '../../../components/ui'
 import { Button } from '../../../components/ui'
-import { useData } from '../hooks/use-data'
-import { prepareTimeSeriesData, formatLargeNumber, generateYAxisTicks, TimePeriod, getAvailableTimePeriods } from '../utils/chart-data'
-import { getFarmingRuns } from '../utils/run-type-filter'
-import { FarmingOnlyIndicator } from './farming-only-indicator'
+import { useData } from '../../data-tracking/hooks/use-data'
+import { prepareTimeSeriesData, getAvailableTimePeriods } from './chart-data'
+import { TimePeriod } from './chart-types'
+import { formatLargeNumber, generateYAxisTicks } from '../../data-tracking/utils/chart-formatters'
+import { getFarmingRuns } from '../../data-tracking/utils/run-type-filter'
+import { FarmingOnlyIndicator } from '../../data-tracking/components/farming-only-indicator'
 
 interface TimeSeriesChartProps {
   metric: 'coins' | 'cells'
