+++ src/routes/charts/coins.tsx
@@ -1,6 +1,6 @@
 import { createFileRoute } from '@tanstack/react-router'
 import { Card, CardContent, CardHeader, CardTitle } from '../../components/ui'
-import { TimeSeriesChart } from '../../features/data-tracking'
+import { TimeSeriesChart } from '../../features/analysis/time-series/time-series-chart'
 
 export const Route = createFileRoute('/charts/coins')({
   component: CoinsChartPage,
